corrected ReplaceInto handling corrected param conversion in estate service

+++ SilverSim/Database.MsSql/Estate/MsSqlEstateService.EstateOwner.cs
@@ -69,7 +69,7 @@ namespace SilverSim.Database.MsSql.Estate
                                 UUI uui = reader.GetUUI("Owner");
                                 if (uui.EqualsGrid(owner))
                                 {
-                                    estates.Add((uint)reader["ID"]);
+                                    estates.Add((uint)(int)reader["ID"]);
                                 }
                             }
                             return estates;
